+++ test/Unit/Ads/PerPlacementLoadManagerTest.ts
@@ -32,7 +32,6 @@ import ConfigurationAuctionPlc from 'json/ConfigurationAuctionPlc.json';
 import 'mocha';
 import * as sinon from 'sinon';
 import { TestFixtures } from 'TestHelpers/TestFixtures';
-import { LoadCalledCounter } from 'Core/Utilities/LoadCalledCounter';
 import { PrivacySDK } from 'Privacy/PrivacySDK';
 import { UserPrivacyManager } from 'Ads/Managers/UserPrivacyManager';
 import { ILoadedCampaign } from 'Ads/Managers/CampaignManager';
@@ -109,16 +108,13 @@ describe('PerPlacementLoadManagerTest', () => {
             let sandbox: sinon.SinonSandbox;
             let loadCampaignStub: sinon.SinonStub;
             let sendReadyEventStub: sinon.SinonStub;
-            let loadCalledKafkaStub: sinon.SinonStub;
 
             beforeEach(() => {
                 sandbox = sinon.createSandbox();
                 placementId = 'premium';
                 loadCampaignStub = sandbox.stub(campaignManager, 'loadCampaign');
                 sendReadyEventStub = sandbox.stub(ads.Listener, 'sendReadyEvent');
-                loadCalledKafkaStub = sandbox.stub(LoadCalledCounter, 'report').callsFake(() => {
-                    return Promise.resolve(<INativeResponse>{});
-                });
+
                 // To silence diagnostic messages
                 sandbox.stub(Diagnostics, 'trigger').callsFake(() => {
                     return Promise.resolve(<INativeResponse>{});
@@ -154,7 +150,6 @@ describe('PerPlacementLoadManagerTest', () => {
                     setPlacementStateStub.callsFake((id: string, state: PlacementState) => {
                         if (placementId === id && state !== PlacementState.WAITING) {
                             const testCampaign = loadManager.getCampaign(placementId);
-                            sinon.assert.called(loadCalledKafkaStub);
                             assert.instanceOf(testCampaign, Campaign, `Campaign with placementID '${placementId}' was not a defined Campaign`);
                             assert.equal(testCampaign, t.expectedCampaign, 'Loaded campaign was not the correct campaign');
                             assert.notEqual(testCampaign, t.unexpectedCampaign, 'Loaded campaign was not the correct campaign');
@@ -194,7 +189,6 @@ describe('PerPlacementLoadManagerTest', () => {
                 loadDict[placementId] = 1;
                 ads.LoadApi.onLoad.trigger(loadDict);
 
-                sinon.assert.called(loadCalledKafkaStub);
                 sinon.assert.notCalled(loadCampaignStub);
                 sinon.assert.notCalled(sendReadyEventStub);
                 sinon.assert.calledWith(<sinon.SinonStub>SDKMetrics.reportMetricEvent, LoadMetric.LoadAuctionRequestBlocked);
@@ -211,7 +205,6 @@ describe('PerPlacementLoadManagerTest', () => {
                 loadDict[placementId] = 1;
                 ads.LoadApi.onLoad.trigger(loadDict);
 
-                sinon.assert.called(loadCalledKafkaStub);
                 sinon.assert.notCalled(loadCampaignStub);
                 sinon.assert.calledWith(sendReadyEventStub, placementId);
                 sinon.assert.calledWith(<sinon.SinonStub>SDKMetrics.reportMetricEvent, LoadMetric.LoadAuctionRequestBlocked);
@@ -232,7 +225,6 @@ describe('PerPlacementLoadManagerTest', () => {
                 placement.setState(PlacementState.READY);
                 placement.setCurrentCampaign(campaign);
 
-                sinon.assert.called(loadCalledKafkaStub);
                 sinon.assert.called(loadCampaignStub);
             });
 
@@ -249,7 +241,6 @@ describe('PerPlacementLoadManagerTest', () => {
                     const placement = adsConfig.getPlacement(placementId);
                     placement.setState(state);
                     return loadManager.initialize().then(() => {
-                        sinon.assert.called(loadCalledKafkaStub);
                         sinon.assert.called(loadCampaignStub);
                     });
                 });