+++ test/Unit/Ads/UserPrivacyManagerTest.ts
@@ -2,7 +2,7 @@ import {
     AgeGateChoice,
     GDPREventAction,
     GDPREventSource,
-    LegalFramework,
+    LegalFramework, OptOutScope,
     UserPrivacyManager
 } from 'Ads/Managers/UserPrivacyManager';
 import { AdsConfiguration } from 'Ads/Models/AdsConfiguration';
@@ -640,7 +640,7 @@ describe('UserPrivacyManagerTest', () => {
 
         it('with proper fields', () => {
             const expectedKafkaObject = {
-                v: 3,
+                v: 4,
                 advertisingId: testAdvertisingId,
                 abGroup: abGroup,
                 layout: layout,
@@ -659,7 +659,8 @@ describe('UserPrivacyManagerTest', () => {
                 permissions: permissions,
                 legalFramework: 'none',
                 agreedOverAgeLimit: 'missing',
-                ageGateSource: 'missing'
+                ageGateSource: 'missing',
+                scope: 'project'
             };
 
             return privacyManager.updateUserPrivacy(permissions, source, action, layout).then(() => {
@@ -730,8 +731,9 @@ describe('UserPrivacyManagerTest', () => {
                     assert.equal(eventData.abGroup, expectedAbGroup);
                     assert.equal(eventData.layout, '');
                     assert.equal(eventData.firstRequest, true);
-                    assert.equal(eventData.v, 3);
+                    assert.equal(eventData.v, 4);
                     assert.equal(eventData.agreedVersion, gamePrivacy.getVersion());
+                    assert.equal(eventData.scope, OptOutScope.PROJECT_SCOPE);
                 });
             });